refactor(catalog): move StructuredValidationError out of api/ into a domain module
StructuredValidationError lived in catalog/api/edit_claims.py, so the
domain service catalog/services/location_paths.py importing it violated
the "api layer is a leaf" boundary.

Move it to a new apps/catalog/exceptions.py, following the convention that
StructuredApiError subclasses live in a domain module near where they're
raised (accounts.auth_errors, core.rate_limits) — not under api/, and not
promoted to core. Repoint all importers (5 api modules, config/api.py, the
service, two tests); no compatibility shim, so the api/ import path is gone.

Drops the corresponding api-leaf BASELINE entry. Remaining baseline: the
soft_delete and export helpers.
