+++ docs/AppBoundaries.md
@@ -65,4 +65,4 @@ Beyond the cross-app tiers, three intra-app invariants are enforced:
 - **The api layer is a leaf** — the domain (models, services, resolve, ingestion, …) must not import back into `api/`. Only catalog has a multi-module `api/` package; the other apps are single `api.py` modules and leaves by construction.
 - **Production code does not import test factories** — `test_factories` is test-only scaffolding and must not enter the runtime path.
 
-Several contracts carry a `# BASELINE` block listing pre-existing violations that predate enforcement: upward imports in `App layer tiers` (e.g. `provenance -> catalog` deferred imports, `core -> accounts` leaks) and domain-logic-stranded-under-`api/` in `The api layer is a leaf` (the `soft_delete`, `export` and `edit_claims` helpers consumed by catalog's domain layer). Each baseline is a burn-down list: fix the underlying import and delete the line. Do not add to one.
+Several contracts carry a `# BASELINE` block listing pre-existing violations that predate enforcement: upward imports in `App layer tiers` (e.g. `provenance -> catalog` deferred imports, `core -> accounts` leaks) and domain-logic-stranded-under-`api/` in `The api layer is a leaf` (the `soft_delete` and `export` helpers consumed by catalog's domain layer). Each baseline is a burn-down list: fix the underlying import and delete the line. Do not add to one.
